+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Adds Kiota bundle package to provide default adapter with registrations setup. [#290](https://github.com/microsoft/kiota-dotnet/issues/290)
 
+## [1.9.12] - 2024-07-30
+
+- Fix non IParasable object serialization.
+- Add basic support for serializing dictionary values in AdditionalData.
+
 ## [1.9.11] - 2024-07-22
 
 - Obsoletes custom decompression handler in favor of native client capabilities.